"Update dependencies and add asset statement
Dependencies for multiple packages across the android and iOS app have been updated to their latest versions, helping to ensure stability, improve performance, and receive the latest bug fixes and features. Furthermore, an asset statement was added to the android app's strings file. The asset statement is part of the digital asset links protocol, which allows the app to show content in android without needing the website's domain name."